requirements.txt
# Django Core
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Configuration
python-decouple==3.8

# Core Python packages
sqlparse==0.5.3
typing_extensions==4.14.1
requests==2.32.3

# ML and Embeddings
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.21.0
huggingface-hub>=0.30.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Vector Database
chromadb>=0.4.0

# Document Processing
PyMuPDF>=1.23.0

# Additional ML dependencies
tokenizers>=0.13.0
tqdm>=4.64.0
filelock>=3.8.0
packaging>=20.0

# Optional but recommended
pillow>=9.0.0
pyyaml>=6.0
